extract some of the models in backend/models.py
