+++ backend/api/invoices/create/services/add.py
@@ -2,7 +2,7 @@ from django.http import HttpRequest, JsonResponse
 from django.shortcuts import render
 from django.views.decorators.http import require_http_methods
 
-from backend.models import InvoiceProduct
+from backend.models_db.invoice import InvoiceProduct
 
 
 @require_http_methods(["POST"])
