+++ backend/api/invoices/delete.py
@@ -5,7 +5,8 @@ from django.urls import resolve
 from django.urls.exceptions import Resolver404
 from django.views.decorators.http import require_http_methods
 
-from backend.models import Invoice, QuotaLimit
+from backend.models import QuotaLimit
+from backend.models_db.invoice import Invoice
 
 
 @require_http_methods(["DELETE"])
