+++ backend/api/receipts/new.py
@@ -5,8 +5,8 @@ from django.shortcuts import render, redirect
 from django.views.decorators.http import require_http_methods
 
 from backend.decorators import quota_usage_check
-from backend.models import Receipt, QuotaUsage
-
+from backend.models import QuotaUsage
+from backend.models_db.receipt import Receipt
 
 @require_http_methods(["POST"])
 @quota_usage_check("receipts-count", api=True, htmx=True)
