+++ backend/api/teams/invites.py
@@ -1,7 +1,8 @@
 from django.http import HttpRequest
 
 from backend.decorators import *
-from backend.models import Notification, Team, TeamInvitation, User
+from backend.models import Team, TeamInvitation, User
+from backend.models_db.notification import Notification
 
 
 def delete_notification(user: User, code: TeamInvitation.code):
