+++ backend/models_db/error.py
@@ -1,0 +1,21 @@
+from django.db import models
+from backend.models import User
+
+class Error(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    error = models.CharField(max_length=250, null=True)
+    error_code = models.CharField(max_length=100, null=True)
+    error_colour = models.CharField(max_length=25, default="danger")
+    date = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return str(self.user_id)
+
+
+class TracebackError(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE, null=True, blank=True)
+    error = models.CharField(max_length=5000, null=True)
+    date = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return str(self.error)
