+++ backend/models_db/notification.py
@@ -1,0 +1,18 @@
+from django.db import models
+
+from backend.models import User
+
+class Notification(models.Model):
+    action_choices = [
+        ("normal", "Normal"),
+        ("modal", "Modal"),
+        ("redirect", "Redirect"),
+    ]
+
+    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="user_notifications")
+    message = models.CharField(max_length=100)
+    action = models.CharField(max_length=10, choices=action_choices, default="normal")
+    action_value = models.CharField(max_length=100, null=True, blank=True)
+    extra_type = models.CharField(max_length=100, null=True, blank=True)
+    extra_value = models.CharField(max_length=100, null=True, blank=True)
+    date = models.DateTimeField(auto_now_add=True)
