+++ backend/models_db/utils.py
@@ -1,0 +1,21 @@
+from django.db import models
+from django.utils.crypto import get_random_string
+
+
+def RandomCode(length=6):
+    return get_random_string(length=length).upper()
+
+
+def RandomAPICode(length=89):
+    return get_random_string(length=length).lower()
+
+
+def USER_OR_ORGANIZATION_CONSTRAINT():
+    return models.CheckConstraint(
+        name=f"%(app_label)s_%(class)s_check_user_or_organization",
+        check=(models.Q(user__isnull=True, organization__isnull=False) | models.Q(user__isnull=False, organization__isnull=True)),
+    )
+
+
+def add_3hrs_from_now():
+    return timezone.now() + timezone.timedelta(hours=3)
