+++ backend/views/core/invoices/create.py
@@ -6,7 +6,9 @@ from django.shortcuts import render, redirect
 from django.views.decorators.http import require_http_methods
 
 from backend.decorators import quota_usage_check
-from backend.models import Invoice, InvoiceItem, Client, InvoiceProduct, QuotaUsage
+from backend.models import QuotaUsage
+from backend.models_db.client import Client
+from backend.models_db.invoice import Invoice, InvoiceItem, InvoiceProduct
 from backend.utils import quota_usage_check_under
