+++ backend/views/core/invoices/manage_access.py
@@ -3,7 +3,8 @@ from django.http import HttpRequest, HttpResponse
 from django.shortcuts import redirect, render
 
 from backend.decorators import quota_usage_check
-from backend.models import Invoice, InvoiceURL, QuotaUsage, QuotaLimit
+from backend.models import QuotaUsage, QuotaLimit
+from backend.models_db.invoice import Invoice, InvoiceURL
 
 
 def manage_access(request: HttpRequest, invoice_id):
